refactor: move getFollowingList function in helper & modify return value
